lib: use ssize_t for read/write return values
read() and write() return ssize_t, storing the result in int
truncates on LP64 where ssize_t is 64-bit. Fix the type in
config_file_read_fd, _sync_issue, get_sysfs_binary,
online_pvid_file_read, online_pvid_file_create,
_btrfs_get_mnt and _sysfs_get_kvdo_value.

+++ lib/device/bcache.c
@@ -444,7 +444,7 @@ static void _sync_destroy(struct io_engine *ioe)
 static bool _sync_issue(struct io_engine *ioe, enum dir d, int di,
                         sector_t sb, sector_t se, void *data, void *context)
 {
-	int rv;
+	ssize_t rv;
 	off_t off;
 	uint64_t where;
 	uint64_t pos = 0;
